src/requirements.txt
# Core framework dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# CrewAI and related dependencies
crewai[tools,agentops]==0.95.0
tavily-python==0.5.0
scrapegraphai-py==1.0.0

# Data processing
pydantic==2.5.0
python-dotenv==1.0.0

# Additional utilities
aiofiles==23.2.1
jinja2==3.1.2

# Optional: For better async support
asyncio-mqtt==0.13.0